+++ frontend/src/routes/_authenticated/annotation-campaign/$campaignID.phase.$phaseType/spectrogram/$spectrogramID.module.scss
